CDDBAPI/Python/requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
dominate==2.8.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-SQLAlchemy-Session==1.1
greenlet==1.1.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.33
protobuf==3.20.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.41
urllib3==2.0.3
visitor==0.1.3
waitress==2.1.2
Werkzeug==2.2.3
